Added fwrapv and Wstrict-overflow=4 to project files. Corrected integer definition in spoton_misc::database(), dbId must be a qint64.
git-svn-id: https://svn.code.sf.net/p/spot-on/code/branches/0.07@2377 0a995953-517f-4c62-b348-104708d926b2

+++ Common/spot-on-misc.cc
@@ -1664,7 +1664,7 @@ void spoton_misc::correctSettingsContainer(QHash<QString, QVariant> settings)
 QSqlDatabase spoton_misc::database(QString &connectionName)
 {
   QSqlDatabase db;
-  int dbId = -1;
+  qint64 dbId = -1;
 
   s_dbMutex.lock();
   dbId = s_dbId += 1;
